requirements.txt
asgiref==3.4.1
backports-zoneinfo==0.2.1
ca-certificates==2022.4.26
certifi==2021.10.8
django-ajax-datatable==4.4.4
django-phonenumber-field==6.1.0
django-tinymce==3.4.0
django==4.0.4
krb5==1.19.2
libcxx==12.0.0
libedit==3.1.20210910
libffi==3.3
libpq==12.9
mysqlclient==2.1.0
ncurses==6.3
openssl==1.1.1n
pip==21.2.4
psycopg2==2.8.6
pymysql==1.0.2
python==3.8.13
pytz==2021.3
readline==8.1.2
setuptools==61.2.0
sqlite==3.38.3
sqlparse==0.4.1
tk==8.6.11
typing_extensions==4.1.1
wheel==0.37.1
xz==5.2.5
zlib==1.2.12
